pyproject.toml
[tool.poetry]
name = "researcher"
version = "0.1.0"
description = ""
authors = ["Shayan Shafquat <shafquat.shayan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pydantic = "^2.0.0"
sqlalchemy = "^2.0.0"
asyncpg = "^0.27.0"
python-multipart = "^0.0.6"
together = "^0.2.0"
unstructured = "^0.7.0"
pinecone-client = "^2.2.1"
faiss-cpu = "^1.7.4"
openai = "^1.3.3"
sentence-transformers = "^2.2.2"
python-dotenv = "^1.0.0"
streamlit = "^1.28.1"
pymupdf = "^1.23.6"
arxiv = "1.4.8"
aiofiles = "^23.2.1"
tiktoken = "^0.8.0"
huggingface-hub = "^0.26.2"
aiohttp = "^3.9.0"
beautifulsoup4 = "^4.12.0"
nltk = "^3.8.1"
requests = "^2.31.0"
urllib3 = "^2.0.7"
PyPDF2 = "^3.0.1"
transformers = "^4.35.0"
torch = "^2.1.0"
llama-index = "^0.11.23"
llama-index-embeddings-langchain = "^0.2.0"
llama-index-llms-langchain = "0.4.0"
text-generation = "^0.7.0"
langchain = "^0.3.7"
langchain-huggingface = "^0.1.2"
langchain-openai = "^0.2.8"
ragas = "^0.2.5"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pymongo = "^4.5.0"
ragas = "^0.2.5"
tqdm = "^4.66.1"
pandas = "^2.2.3"
text-generation = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
